backend/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
configparser==6.0.0
Django==5.0
django-cors-headers==3.9.0
django-filter==23.5
django-ipware==6.0.3
django-rest-authemail==2.1.7
djangorestframework==3.14.0
Faker==22.5.1
idna==3.6
Markdown==3.5.2
python-dateutil==2.8.2
python-ipware==2.0.1
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.1.0
